requirements.txt
# Bug Bounty Security Scanner - Dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
pyyaml>=6.0
rich>=12.0.0
